Modificaciones Lucas

+++ Backend/src/class/usuario/usuario-model.js
@@ -44,10 +44,10 @@ const UsuarioModelo = sequelize.define('usuario', {
 		type: Sequelize.INTEGER
 	},
 	nroCelularUsuario: {
-		type: Sequelize.INTEGER
+		type: Sequelize.BIGINT(10)
 	},
 	nroTelefonoUsuario: {
-		type: Sequelize.BIGINT(15)
+		type: Sequelize.BIGINT(10)
 	},
 	  
   }, {
